# Lab2/ark-pzpi-23-3-osychenko-illia-lab2/docs/swagger.yaml
basePath: /api
definitions:
...
  model.Bus:
    properties:
      capacity:
        example: 50
        type: integer
      fuel_consumption_per_100km:
        example: 12.5
        type: number
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      model:
        example: Mercedes Sprinter
        type: string
      registration_number:
        example: AA1234BB
        type: string
    type: object
  model.PassengerEvent:
    properties:
      device_local_id:
        example: 123
        type: integer
      event_type:
        enum:
        - board
        - alight
        example: board
        type: string
      id:
        example: 1
        type: integer
      is_synced:
        example: true
        type: boolean
      latitude:
        example: 49.9935
        type: number
      longitude:
        example: 36.2304
        type: number
      passenger_count_after:
        example: 25
        type: integer
      timestamp:
        example: "2023-12-15T08:15:00Z"
        type: string
      trip_id:
        example: 1
        type: integer
    type: object
...
  model.Route:
    properties:
      base_price:
        example: 250
        type: number
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      destination_city:
        example: Київ
        type: string
      distance_km:
        example: 480.5
        type: number
      driver_cost_per_trip:
        example: 800
        type: number
      estimated_duration_minutes:
        example: 360
        type: integer
      fuel_cost_per_km:
        example: 2.5
        type: number
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      origin_city:
        example: Харків
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  model.Trip:
    properties:
      actual_arrival:
        example: "2023-12-15T14:05:00Z"
        type: string
      actual_departure:
        example: "2023-12-15T08:05:00Z"
        type: string
      bus:
        $ref: '#/definitions/model.Bus'
      bus_id:
        example: 1
        type: integer
      current_passengers:
        example: 35
        type: integer
      driver_name:
        example: Петро Петренко
        type: string
      id:
        example: 1
        type: integer
      route:
        $ref: '#/definitions/model.Route'
      route_id:
        example: 1
        type: integer
      scheduled_departure:
        example: "2023-12-15T08:00:00Z"
        type: string
      status:
        enum:
        - scheduled
        - in_progress
        - completed
        - cancelled
        example: completed
        type: string
    type: object
  model.TripAnalytics:
    properties:
      avg_occupancy_rate:
        example: 0.75
        type: number
      calculated_at:
        example: "2023-12-15T20:00:00Z"
        type: string
      driver_cost:
        example: 800
        type: number
      fuel_cost:
        example: 1200.25
        type: number
      id:
        example: 1
        type: integer
      max_passengers:
        example: 45
        type: integer
      other_costs:
        example: 150
        type: number
      profit:
        example: 1399.75
        type: number
      profitability_percent:
        example: 39.43
        type: number
      revenue:
        example: 3550
        type: number
      total_passengers:
        example: 142
        type: integer
      trip_id:
        example: 1
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@busoptima.ua
    name: API Support
  description: API для системи оптимізації автобусних перевезень BusOptima
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BusOptima API
  version: "1.0"
paths:
...
  /trips:
    get:
      consumes:
      - application/json
      description: Повертає список всіх рейсів з можливістю фільтрації
      parameters:
      - description: ID маршруту для фільтрації
        in: query
        name: route_id
        type: integer
      - description: Статус рейсу (scheduled, in_progress, completed, cancelled)
        in: query
        name: status
        type: string
      - description: Дата початку періоду (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Дата кінця періоду (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Trip'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати список рейсів
      tags:
      - Trips
    post:
      consumes:
      - application/json
      description: Створює новий рейс в системі
      parameters:
      - description: Дані рейсу
        in: body
        name: trip
        required: true
        schema:
          $ref: '#/definitions/model.Trip'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Trip'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Створити новий рейс
      tags:
      - Trips
  /trips/{id}:
    get:
      consumes:
      - application/json
      description: Повертає рейс за вказаним ідентифікатором
      parameters:
      - description: ID рейсу
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Trip'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати рейс за ID
      tags:
      - Trips
    put:
      consumes:
      - application/json
      description: Оновлює існуючий рейс за ID
      parameters:
      - description: ID рейсу
        in: path
        name: id
        required: true
        type: integer
      - description: Оновлені дані рейсу
        in: body
        name: trip
        required: true
        schema:
          $ref: '#/definitions/model.Trip'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Trip'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Оновити рейс
      tags:
      - Trips
  /trips/{id}/analytics:
    get:
      consumes:
      - application/json
      description: Повертає детальну аналітику для конкретного рейсу
      parameters:
      - description: ID рейсу
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TripAnalytics'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати аналітику рейсу
      tags:
      - Trips
  /trips/{id}/events:
    get:
      consumes:
      - application/json
      description: Повертає список подій входу/виходу пасажирів для конкретного рейсу
      parameters:
      - description: ID рейсу
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PassengerEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати події пасажирів рейсу
      tags:
      - Trips
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"